🔖 rename >> 컨밴션에 맞는 파일, 폴더이름 수정
