+++ AdminTogglePage/src/libs/styledComponents/GlobalStyles.jsx
@@ -1,6 +1,6 @@
 import { createGlobalStyle } from "styled-components"
 
-import { BREAK_POINT, COLOR, FONT_SIZE } from "./reference-tokens"
+import { BREAK_POINT, COLOR, FONT_SIZE } from "./referenceTokens"
 
 const GlobalStyles = createGlobalStyle`
   *{
