--- AdminTogglePage/src/libs/styledComponents/reference-tokens/index.js
@@ -1,3 +1,0 @@
-export * from "./break-point";
-export * from "./color";
-export * from "./font-size";
