+++ AdminTogglePage/src/main.jsx
@@ -2,7 +2,7 @@ import React from "react"
 import ReactDOM from "react-dom/client"
 import { RouterProvider } from "react-router-dom"
 import router from "./libs/router/router.jsx"
-import GlobalStyles from "./libs/styledComponents/global-styles.jsx"
+import GlobalStyles from "./libs/styledComponents/GlobalStyles.jsx"
 
 ReactDOM.createRoot(document.getElementById("root")).render(
   <React.StrictMode>
